# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Locally Imagined backend
    description: Serves all frontend requests
    version: ""
host: localhost:80
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /posts/artistposts/{page}:
        get:
            tags:
                - postings
            summary: get_artist_post_page postings
            operationId: postings#get_artist_post_page
            parameters:
                - name: userID
                  in: query
                  description: User ID to get posts for
                  required: true
                  type: string
                - name: page
                  in: path
                  description: Page to get posts for
                  required: true
                  type: integer
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/PostResponse'
                "401":
                    description: Unauthorized response.
                    schema:
                        $ref: '#/definitions/PostingsGetArtistPostPageUnauthorizedResponseBody'
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/PostingsGetArtistPostPageInternalResponseBody'
            schemes:
                - http
    /posts/create:
        post:
            tags:
                - postings
            summary: create_post postings
            operationId: postings#create_post
            parameters:
                - name: Authorization
                  in: header
                  description: jwt used for auth
                  required: true
                  type: string
                - name: create_post_request_body
                  in: body
                  description: Post info
                  required: true
                  schema:
                    $ref: '#/definitions/PostingsCreatePostRequestBody'
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/PostingsCreatePostResponseBody'
                        required:
                            - title
                            - description
                            - price
                            - imageIDs
                            - postID
                            - medium
                            - uploadDate
                            - sold
                            - deliverytype
                            - userID
                            - username
                "401":
                    description: Unauthorized response.
                    schema:
                        $ref: '#/definitions/PostingsCreatePostUnauthorizedResponseBody'
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/PostingsCreatePostInternalResponseBody'
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
    /posts/delete/{postID}:
        delete:
            tags:
                - postings
            summary: delete_post postings
            operationId: postings#delete_post
            parameters:
                - name: postID
                  in: path
                  description: Post to delete
                  required: true
                  type: string
                - name: Authorization
                  in: header
                  description: jwt used for auth
                  required: true
                  type: string
            responses:
                "204":
                    description: No Content response.
                "401":
                    description: Unauthorized response.
                    schema:
                        $ref: '#/definitions/PostingsDeletePostUnauthorizedResponseBody'
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/PostingsDeletePostInternalResponseBody'
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
    /posts/edit/{postID}:
        put:
            tags:
                - postings
            summary: edit_post postings
            operationId: postings#edit_post
            parameters:
                - name: title
                  in: query
                  description: Post title
                  required: false
                  type: string
                - name: description
                  in: query
                  description: Post description
                  required: false
                  type: string
                - name: price
                  in: query
                  description: Post price
                  required: false
                  type: string
                - name: content
                  in: query
                  description: Image content
                  required: false
                  type: string
                - name: medium
                  in: query
                  description: Art type
                  required: false
                  type: string
                - name: sold
                  in: query
                  description: is sold
                  required: false
                  type: boolean
                - name: deliverytype
                  in: query
                  description: Delivery type
                  required: false
                  type: string
                - name: imageID
                  in: query
                  description: Image ID
                  required: false
                  type: string
                - name: postID
                  in: path
                  description: Post ID
                  required: true
                  type: string
                - name: Authorization
                  in: header
                  description: jwt used for auth
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/PostingsEditPostResponseBody'
                        required:
                            - title
                            - description
                            - price
                            - imageIDs
                            - postID
                            - medium
                            - uploadDate
                            - sold
                            - deliverytype
                            - userID
                            - username
                "401":
                    description: Unauthorized response.
                    schema:
                        $ref: '#/definitions/PostingsEditPostUnauthorizedResponseBody'
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/PostingsEditPostInternalResponseBody'
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
    /posts/getimages/{postID}:
        get:
            tags:
                - postings
            summary: get_images_for_post postings
            operationId: postings#get_images_for_post
            parameters:
                - name: postID
                  in: path
                  description: Post to get images for
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: array
                        items:
                            type: string
                            example: Molestiae veritatis natus quae expedita similique.
                "401":
                    description: Unauthorized response.
                    schema:
                        $ref: '#/definitions/PostingsGetImagesForPostUnauthorizedResponseBody'
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/PostingsGetImagesForPostInternalResponseBody'
            schemes:
                - http
    /posts/getpage/{page}:
        get:
            tags:
                - postings
            summary: get_post_page postings
            operationId: postings#get_post_page
            parameters:
                - name: page
                  in: path
                  description: Page to get posts for
                  required: true
                  type: integer
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/PostResponse'
                "401":
                    description: Unauthorized response.
                    schema:
                        $ref: '#/definitions/PostingsGetPostPageUnauthorizedResponseBody'
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/PostingsGetPostPageInternalResponseBody'
            schemes:
                - http
    /posts/getpagefiltered/{page}:
        get:
            tags:
                - postings
            summary: get_post_page_filtered postings
            operationId: postings#get_post_page_filtered
            parameters:
                - name: keyword
                  in: query
                  description: Search bar keyword to search for in title and description
                  required: false
                  type: string
                - name: startDate
                  in: query
                  description: Filter attribute to see posts after given date
                  required: false
                  type: string
                - name: endDate
                  in: query
                  description: Filter attribute to see posts before given date
                  required: false
                  type: string
                - name: medium
                  in: query
                  description: Filter attribute to see posts with given medium type
                  required: false
                  type: string
                - name: page
                  in: path
                  description: Page to get posts for
                  required: true
                  type: integer
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/PostResponse'
                "401":
                    description: Unauthorized response.
                    schema:
                        $ref: '#/definitions/PostingsGetPostPageFilteredUnauthorizedResponseBody'
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/PostingsGetPostPageFilteredInternalResponseBody'
            schemes:
                - http
...
definitions:
...
    PostResponse:
        title: PostResponse
        type: object
        properties:
            deliverytype:
                type: string
                description: Delivery type
                example: Repellat esse odit corporis officia quod.
            description:
                type: string
                description: Post description
                example: Nihil dolore consequatur molestias non consectetur et.
            imageIDs:
                type: array
                items:
                    type: string
                    example: Cupiditate autem.
                description: Image ID
                example:
                    - Unde beatae sint enim alias delectus odio.
                    - Quaerat ut eveniet officiis eum.
                    - Consequuntur veniam architecto aut nobis vitae voluptatem.
            medium:
                type: string
                description: Art type
                example: Illum debitis accusantium.
            postID:
                type: string
                description: Post ID
                example: Autem id eligendi at.
            price:
                type: string
                description: Post price
                example: Dolore dolorum illum ab rerum nobis.
            sold:
                type: boolean
                description: is sold
                example: false
            title:
                type: string
                description: Post title
                example: Eum quo recusandae.
            uploadDate:
                type: string
                description: Upload Date
                example: Est quisquam aliquam sed sit.
            userID:
                type: string
                description: User id associated with post
                example: Rerum consequatur excepturi dicta fugiat.
            username:
                type: string
                description: Username associated with post
                example: Aut qui laudantium culpa voluptas quos magni.
        description: Describes a post response
        example:
            deliverytype: Doloribus fugit rem.
            description: Autem adipisci impedit molestiae in officiis quod.
            imageIDs:
                - Rerum rerum itaque.
                - Minus minima hic explicabo.
                - Aut nihil dolor nisi perspiciatis.
                - Repellendus expedita iste et ducimus id.
            medium: Qui esse iusto iste.
            postID: Vero aut ex dolore.
            price: Harum vel hic iste eos.
            sold: false
            title: Eveniet totam et cum dicta magnam sunt.
            uploadDate: Modi soluta ipsam iusto omnis cupiditate.
            userID: Est vel.
            username: Maxime dolores.
        required:
            - title
            - description
            - price
            - imageIDs
            - postID
            - medium
            - uploadDate
            - sold
            - deliverytype
            - username
            - userID
...
    PostingsCreatePostResponseBody:
        title: PostingsCreatePostResponseBody
        $ref: '#/definitions/PostResponseResponseBody'
...
    PostingsEditPostResponseBody:
        title: PostingsEditPostResponseBody
        $ref: '#/definitions/PostResponseResponseBody'
...
securityDefinitions:
    jwt_header_Authorization:
        type: apiKey
        description: Secures endpoint by requiring a valid JWT token retrieved via the login endpoint.
        name: Authorization
        in: header
    login_header_Authorization:
        type: basic
        description: Basic authentication used to authenticate security principal during login
    signup_header_Authorization:
        type: basic
        description: Basic authentication used to authenticate security principal during signup